#72-Capturing frames rate is no more ok
Bug is now solved. his led to the development of many new modules and classes to alleviate the misfunctionning of Windows (and of Linux as well) as a non-RTOS (Real-Time Operating System).
This was an interesting task.
Completed, Tested.
Not all artifacts have been solved, and will not be easily. They might not be solvable either. Remember: WIndows is NOT an RTOS. Neither is Linux also.